Move grid_one_variable() to its own module; functionize parts.
